Feat: jpa 테스트코드 추가

+++ grid/src/main/java/org/highfives/grid/vacation/command/repository/VacationPolicyRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VacationPolicyRepository extends JpaRepository<VacationPolicy, Long> {
+public interface VacationPolicyRepository extends JpaRepository<VacationPolicy, Integer> {
 }
