+++ grid/src/main/java/org/highfives/grid/vacation/command/repository/VacationTypeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VacationTypeRepository extends JpaRepository<VacationType, Long> {
+public interface VacationTypeRepository extends JpaRepository<VacationType, Integer> {
 
 }
